Clean-up:  - remove server mocking  - remove sessionUserActionSender  - remove sessionTraitsSender
