+++ src/gravity-client/NopGravityClient.ts
@@ -1,8 +1,6 @@
-import { SessionTraits, SessionUserAction } from '../types'
+import { SessionTraits, SessionUserAction, AddSessionUserActionsResponse, IdentifySessionResponse } from '../types'
 import { AbstractGravityClient } from './AbstractGravityClient'
 import { IGravityClient } from './IGravityClient'
-import { AddSessionUserActionsResponse } from '../user-action/sessionUserActionSender'
-import { IdentifySessionResponse } from '../session-trait/sessionTraitSender'
 
 export default class NopGravityClient extends AbstractGravityClient implements IGravityClient {
     async handleSessionUserActions(
