--- src/mocks/server.ts
@@ -1,4 +1,0 @@
-import { setupServer } from 'msw/node'
-import { handlers } from './handlers'
-
-export const server = setupServer(...handlers)
